third_party/WebKit/LayoutTests/dom/attr/direction-attribute-set-and-cleared-expected.txt

This tests that setting the dir attribute to rtl and clearing it defaults to ltr.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS window.getComputedStyle(document.body).direction is "ltr"
